make synchronization better

+++ src/main/java/com/openSolutions/currencyJournal/service/interfaces/CbrSyncService.java
@@ -3,6 +3,6 @@ package com.openSolutions.currencyJournal.service.interfaces;
 import com.openSolutions.currencyJournal.domain.dto.response.StatusResponse;
 
 public interface CbrSyncService {
-    int synchronizeWithCbr();
+    long synchronizeWithCbr();
     StatusResponse getStatusInfo();
 }
